requirements.txt
alpha_vantage
fredapi
lxml
matplotlib
nltk
pandas
pymongo
python-dotenv
seaborn
tweepy
yfinance
